requirements.txt
PyExifTool==0.4.13